Refactor InputType values for clarity and consistency
