+++ src/web/platform/pillars/strength.py
@@ -3,7 +3,7 @@ from src.web.platform.core.definitions import (
     InputDefinition,
     PersonaDefinition,
 )
-from src.web.platform.core.types import FormattingType, Icon, InputType
+from src.web.platform.core.enums import FormattingType, Icon, InputType
 
 vanguard_persona = PersonaDefinition(
     public_name="Vanguarda",
